C57 compiles/runs.

+++ chapter-57-ai/src/ai/approach_ai_system.rs
@@ -26,15 +26,15 @@ impl<'a> System<'a> for ApproachAI {
         {
             turn_done.push(entity);
             let path = rltk::a_star_search(
-                map.xy_idx(pos.x, pos.y) as i32,
-                map.xy_idx(approach.idx % map.width, approach.idx / map.width) as i32,
+                map.xy_idx(pos.x, pos.y),
+                map.xy_idx(approach.idx % map.width, approach.idx / map.width),
                 &mut *map
             );
             if path.success && path.steps.len()>1 {
                 let mut idx = map.xy_idx(pos.x, pos.y);
                 map.blocked[idx] = false;
-                pos.x = path.steps[1] % map.width;
-                pos.y = path.steps[1] / map.width;
+                pos.x = path.steps[1] as i32 % map.width;
+                pos.y = path.steps[1] as i32 / map.width;
                 entity_moved.insert(entity, EntityMoved{}).expect("Unable to insert marker");
                 idx = map.xy_idx(pos.x, pos.y);
                 map.blocked[idx] = true;
