+++ chapter-57-ai/src/ai/chase_ai_system.rs
@@ -44,15 +44,15 @@ impl<'a> System<'a> for ChaseAI {
             turn_done.push(entity);
             let target_pos = targets[&entity];
             let path = rltk::a_star_search(
-                map.xy_idx(pos.x, pos.y) as i32,
-                map.xy_idx(target_pos.0, target_pos.1) as i32,
+                map.xy_idx(pos.x, pos.y),
+                map.xy_idx(target_pos.0, target_pos.1),
                 &mut *map
             );
             if path.success && path.steps.len()>1 && path.steps.len()<15 {
                 let mut idx = map.xy_idx(pos.x, pos.y);
                 map.blocked[idx] = false;
-                pos.x = path.steps[1] % map.width;
-                pos.y = path.steps[1] / map.width;
+                pos.x = path.steps[1] as i32 % map.width;
+                pos.y = path.steps[1] as i32 / map.width;
                 entity_moved.insert(entity, EntityMoved{}).expect("Unable to insert marker");
                 idx = map.xy_idx(pos.x, pos.y);
                 map.blocked[idx] = true;
