+++ chapter-59-caverns2/src/ai/chase_ai_system.rs
@@ -44,8 +44,8 @@ impl<'a> System<'a> for ChaseAI {
             turn_done.push(entity);
             let target_pos = targets[&entity];
             let path = rltk::a_star_search(
-                map.xy_idx(pos.x, pos.y) as i32,
-                map.xy_idx(target_pos.0, target_pos.1) as i32,
+                map.xy_idx(pos.x, pos.y),
+                map.xy_idx(target_pos.0, target_pos.1),
                 &mut *map
             );
             if path.success && path.steps.len()>1 && path.steps.len()<15 {
