+++ chapter-63-effects/src/ai/approach_ai_system.rs
@@ -25,8 +25,8 @@ impl<'a> System<'a> for ApproachAI {
         {
             turn_done.push(entity);
             let path = rltk::a_star_search(
-                map.xy_idx(pos.x, pos.y) as i32,
-                map.xy_idx(approach.idx % map.width, approach.idx / map.width) as i32,
+                map.xy_idx(pos.x, pos.y),
+                map.xy_idx(approach.idx % map.width, approach.idx / map.width),
                 &mut *map
             );
             if path.success && path.steps.len()>1 {
